Include more paths
Getting an error from the aws cli: Unkowon encoding: idna
Upon investigation, error only happens when calling the aws cli. Further
research looks like aws includes an "encode()" call that uses some files
for encoding. Not sure the details but used "tree | grep encoding" to
find the files in the uncompressed file system which has the word
encoding in it (I'm presuming the error is saying it's missing a file)